Refactor transaction and withdrawal handling; rename WITHDRAW to WITHDRAWAL in TransactionType enum, change accountNumberHolder type to long in Withdrawal model, and add WithdrawalDTO for withdrawal data transfer. Implement HomePageController and repositories for Withdrawal and Transaction.

+++ src/main/java/vn/project/group_lottery/model/Withdrawal.java
@@ -25,7 +25,7 @@ public class Withdrawal {
     @Enumerated(EnumType.STRING)
     private BankName bankName;
 
-    private int accountNumberHolder;
+    private long accountNumberHolder;
 
     private String bankNameHolder;
 }
